+++ packages/blocks/src/blocks/core-list-item.tsx
@@ -5,6 +5,10 @@ import {
 	getStylesFromAttributes,
 } from '@snapwp/core';
 import { Parse } from '@snapwp/next';
+import type {
+	CoreListItem as CoreListItemType,
+	CoreListItemProps,
+} from '@snapwp/types';
 
 /**
  * Renders the core/list-item block.
@@ -16,11 +20,11 @@ import { Parse } from '@snapwp/next';
  *
  * @return The rendered block.
  */
-export default function CoreListItem( {
+const CoreListItem: CoreListItemType = ( {
 	attributes,
 	renderedHtml,
 	children,
-}: any ) {
+}: CoreListItemProps ) => {
 	const { content, style } = attributes || {};
 
 	const styleObject = getStylesFromAttributes( { style } );
@@ -41,4 +45,6 @@ export default function CoreListItem( {
 			{ children }
 		</li>
 	);
-}
+};
+
+export default CoreListItem;