+++ packages/next/src/template-renderer/index.tsx
@@ -1,10 +1,10 @@
 import React, { type JSX } from 'react';
 import { headers } from 'next/headers';
-import { type BlockData } from '@snapwp/core';
 import { QueryEngine } from '@snapwp/query';
 import { TemplateHead } from './template-head';
 import { TemplateScripts } from './template-scripts';
 import Script from 'next/script';
+import type { BlockData } from '@snapwp/types';
 
 export type TemplateRendererProps = {
 	getTemplateData?: ( typeof QueryEngine )[ 'getTemplateData' ];