+++ packages/query/src/utils/parse-template.ts
@@ -1,12 +1,11 @@
 import {
 	Logger,
-	TemplateParseError,
-	type BlockData,
 	type EnqueuedScriptProps,
-	type ScriptModuleProps,
 	type StyleSheetProps,
-	type TemplateData,
+	type ScriptModuleProps,
+	TemplateParseError,
 } from '@snapwp/core';
+import type { BlockData } from '@snapwp/types';
 import type { ApolloQueryResult } from '@apollo/client';
 import type { GetCurrentTemplateQuery } from '@graphqlTypes/graphql';
 
@@ -23,7 +22,7 @@ export default function parseQueryResult(
 	queryData: ApolloQueryResult< GetCurrentTemplateQuery >,
 	wordpressUrl: string,
 	uri: string
-): TemplateData {
+) {
 	if ( queryData.errors?.length ) {
 		queryData.errors?.forEach( ( error ) => {
 			Logger.error(
@@ -34,7 +33,11 @@ export default function parseQueryResult(
 		} );
 	}
 
-	if ( ! queryData.data && queryData.errors?.length ) {
+	// If there is no data or templateByUri in the query data and there are errors, throw an error.
+	if (
+		( ! queryData.data || ! queryData.data.templateByUri ) &&
+		queryData.errors?.length
+	) {
 		throw new TemplateParseError(
 			`Error fetching template data for uri: ${ uri }`
 		);