+++ packages/types/src/blocks/props/core-columns.ts
@@ -1,0 +1,13 @@
+import type { PropsWithChildren } from 'react';
+import type { BaseAttributes, BaseProps } from '..';
+
+export type CoreColumnsAttributes = BaseAttributes & {
+	style?: string;
+	cssClassName?: string;
+};
+
+export type CoreColumnsProps = PropsWithChildren<
+	BaseProps< CoreColumnsAttributes >
+>;
+
+export type CoreColumns = React.ComponentType< CoreColumnsProps >;