+++ packages/types/src/blocks/props/core-details.ts
@@ -1,0 +1,14 @@
+import type { PropsWithChildren } from 'react';
+import type { BaseAttributes, BaseProps } from '../base';
+
+export type CoreDetailsAttributes = BaseAttributes & {
+	style?: string;
+	showContent: boolean;
+	summary?: string;
+};
+
+export type CoreDetailsProps = PropsWithChildren<
+	BaseProps< CoreDetailsAttributes >
+>;
+
+export type CoreDetails = React.ComponentType< CoreDetailsProps >;