+++ packages/types/src/blocks/props/core-html.ts
@@ -1,0 +1,5 @@
+import type { BaseProps } from '../base';
+
+export type CoreHtmlProps = BaseProps< {} >;
+
+export type CoreHtml = React.ComponentType< CoreHtmlProps >;