+++ packages/types/src/blocks/props/core-separator.ts
@@ -1,0 +1,5 @@
+import type { BaseProps } from '../base';
+
+export type CoreSeparatorProps = BaseProps< {} >;
+
+export type CoreSeparator = React.ComponentType< CoreSeparatorProps >;